God knows where I last ended. But still continued. This time sober though!!!
Some sound of mind, sober, new stuff should be:
- works on phone, somewhat.
- shortened wait for freaking doubleclick-things detection threshold
- **export file** successfully exports the data to a json file in the Android documents folder. Yay, this was the main mission of continuing
- **ALAS:** clearing the data file produces a 381 day accounting error
- Continued with my DEARME.md file, proudishly, not egotistical and self-centered, just some inner-confidence :)

Till later, cheers. Branching time too, methinks. Almost fuck shit up baaaaadly, as is the norm - of old behaviour!

+++ src/app/export/export.component.ts
@@ -1,11 +1,11 @@
 import { Component, inject } from "@angular/core";
 import { Filesystem, Directory, Encoding } from "@capacitor/filesystem";
-import {} from "@angular/core";
 import { BeanServiceService } from "../bean-service.service";
 import { OmniscientTransaction } from "../transaction";
 import { JsonPipe } from "@angular/common";
 import { StatusControlService } from "../status-control.service";
 import { LedgerViewComponent } from "../ledger-view/ledger-view.component";
+// import { AmodalComponent } from "../amodal/amodal.component";
 
 @Component({
   selector: "app-export",
@@ -24,6 +24,9 @@ export class ExportComponent {
   constructor() {}
 
   ngOnInit() {
+    // make a backbutton
+    this.statusControl.activeLeftBtn.set("History");
+
     console.log("In oninit of export");
     this.getLastSavedDate();
 
@@ -92,6 +95,7 @@ export class ExportComponent {
 
   getPlatform(): string {
     console.log("platform using:");
+    
     let detectedPlatform: string = "unknown";
     if (typeof window !== "undefined" && (window as any).Capacitor) {
       detectedPlatform = (window as any).Capacitor.getPlatform();
@@ -130,7 +134,15 @@ export class ExportComponent {
   }
 
   async clearFile() {
-    localStorage.clear()
+    const confirmed = confirm("Are you sure you want to clear the file? This cannot be undone.");
+    if (!confirmed) {
+      console.log("File clear cancelled by user.");
+      return;
+    }
+    
+    this.beanService.balanceData.lastUpdated = new Date().getTime();
+    this.beanService.balance = 0;
+    localStorage.clear();
   }
 
   async permissionit(): Promise<boolean> {
